web/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
anthropic==0.40.0
openai==1.50.0
httpx==0.27.0
yt-dlp>=2024.1.0
python-multipart==0.0.9
